refactor: reuse existing V4 mutation implementations
