+++ grovedb-version/src/version/v4.rs
@@ -436,9 +436,9 @@ pub const GROVE_V4: GroveVersion = GroveVersion {
             },
             insert: GroveDBOperationsInsertVersions {
                 insert: 0,
-                // v2: automatic backward-reference maintenance, observing old
+                // v1: automatic backward-reference maintenance, observing old
                 // values in Merks retained for the write.
-                insert_on_transaction: 2,
+                insert_on_transaction: 1,
                 // v2: a directly inserted Reference binds the value hash of its
                 // terminal's STORED bytes (wrapper included for a NonCounted
                 // terminal), matching what the batch reference resolver has
@@ -476,9 +476,9 @@ pub const GROVE_V4: GroveVersion = GroveVersion {
                 // with the child's tree type (issue #686). v0 (GROVE_V1..V3)
                 // keeps the legacy reopen byte-for-byte for replay
                 // compatibility.
-                // v3: automatic backward-reference maintenance with cached
+                // v2: automatic backward-reference maintenance with cached
                 // old-value observation. Skip retains the v1 route.
-                delete_internal_on_transaction: 3,
+                delete_internal_on_transaction: 2,
                 average_case_delete_operation_for_delete: 0,
                 worst_case_delete_operation_for_delete: 0,
             },
