+++ grovedb/src/operations/insert/add_element_on_transaction/mod.rs
@@ -43,7 +43,6 @@
 mod v0;
 mod v1;
 mod v2;
-mod v3;
 
 use grovedb_costs::{CostResult, CostsExt, OperationCost};
 use grovedb_merk::Merk;
