+++ grovedb/src/operations/insert/add_element_on_transaction/v2.rs
@@ -90,6 +90,36 @@ impl GroveDb {
                 grove_version
             )
         );
+        cost_return_on_error!(
+            &mut cost,
+            self.add_element_to_cached_merk_v2(
+                &mut subtree_to_insert_into,
+                path,
+                key,
+                element,
+                options,
+                transaction,
+                batch,
+                grove_version,
+            )
+        );
+        Ok(subtree_to_insert_into).wrap_with_cost(cost)
+    }
+
+    /// V4 insertion mechanics applied to an already prepared Merk.
+    pub(crate) fn add_element_to_cached_merk_v2<'db, B: AsRef<[u8]>>(
+        &'db self,
+        subtree_to_insert_into: &mut Merk<PrefixedRocksDbTransactionContext<'db>>,
+        path: SubtreePath<B>,
+        key: &[u8],
+        element: Element,
+        options: InsertOptions,
+        transaction: &'db Transaction,
+        batch: &'db StorageBatch,
+        grove_version: &GroveVersion,
+    ) -> CostResult<(), Error> {
+        let mut cost = OperationCost::default();
+
         // if we don't allow a tree override then we should check
 
         if options.checks_for_override() {
@@ -193,7 +223,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_reference(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         referenced_element_value_hash,
                         Some(options.as_merk_options()),
@@ -224,7 +254,7 @@ impl GroveDb {
                     cost_return_on_error_into!(
                         &mut cost,
                         element.insert_subtree(
-                            &mut subtree_to_insert_into,
+                            subtree_to_insert_into,
                             key,
                             NULL_HASH,
                             Some(options.as_merk_options()),
@@ -240,7 +270,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         grovedb_commitment_tree::EMPTY_COMMITMENT_TREE_STATE_ROOT,
                         Some(options.as_merk_options()),
@@ -272,7 +302,7 @@ impl GroveDb {
                 let already_exists = cost_return_on_error_into!(
                     &mut cost,
                     element.element_at_key_already_exists(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         grove_version,
                     )
@@ -293,7 +323,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         grovedb_private_document_store::empty_private_document_store_state_root(
                             *entry_size,
@@ -312,7 +342,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         NULL_HASH,
                         Some(options.as_merk_options()),
@@ -343,7 +373,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         Some(options.as_merk_options()),
                         grove_version
@@ -364,191 +394,189 @@ impl GroveDb {
                 .wrap_with_cost(cost);
             }
             Element::ProvableCountIndexedTree(primary, secondary, count_value, _) => {
-                let (primary_root_hash, secondary_root_hash) = if primary.is_none()
-                    && secondary.is_none()
-                    && *count_value == 0
-                {
-                    // Empty cidx: both root keys absent AND count
-                    // is zero. NULL_HASH for both Merks.
-                    (NULL_HASH, NULL_HASH)
-                } else {
-                    // Non-empty cidx: REQUIRE both root_keys to be
-                    // Some(_) AND validate them against on-disk
-                    // state. Reject partially-initialized claims
-                    // explicitly:
-                    //   - (None, None, count > 0): a cidx claiming
-                    //     entries but with no roots — would persist
-                    //     a count_value disconnected from any real
-                    //     index content.
-                    //   - (Some, None, _) / (None, Some, _): only
-                    //     one of the two Merks claimed; would
-                    //     persist asymmetric roots that fail H1-A
-                    //     reconstruction.
-                    if primary.is_none() || secondary.is_none() {
-                        return Err(Error::InvalidInput(
-                            "CountIndexedTree direct insertion: non-empty cidx must \
+                let (primary_root_hash, secondary_root_hash) =
+                    if primary.is_none() && secondary.is_none() && *count_value == 0 {
+                        // Empty cidx: both root keys absent AND count
+                        // is zero. NULL_HASH for both Merks.
+                        (NULL_HASH, NULL_HASH)
+                    } else {
+                        // Non-empty cidx: REQUIRE both root_keys to be
+                        // Some(_) AND validate them against on-disk
+                        // state. Reject partially-initialized claims
+                        // explicitly:
+                        //   - (None, None, count > 0): a cidx claiming
+                        //     entries but with no roots — would persist
+                        //     a count_value disconnected from any real
+                        //     index content.
+                        //   - (Some, None, _) / (None, Some, _): only
+                        //     one of the two Merks claimed; would
+                        //     persist asymmetric roots that fail H1-A
+                        //     reconstruction.
+                        if primary.is_none() || secondary.is_none() {
+                            return Err(Error::InvalidInput(
+                                "CountIndexedTree direct insertion: non-empty cidx must \
                                  have BOTH primary_root_key and secondary_root_key set \
                                  to Some(_); partial state (one None, one Some, or \
                                  count>0 with no roots) is not permitted",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    // Both roots are Some(_); open and verify they
-                    // match the on-disk state. Mismatch ⇒ the
-                    // element bytes would diverge from on-disk
-                    // state; refuse rather than persist an
-                    // inconsistent root_hash chain.
-                    let child_path_owned = path.derive_owned_with_child(key.to_vec());
-                    let child_path = SubtreePath::from(&child_path_owned);
-                    let primary_merk = cost_return_on_error!(
-                        &mut cost,
-                        self.open_transactional_merk_at_path(
-                            child_path.clone(),
-                            transaction,
-                            Some(batch),
-                            grove_version,
-                        )
-                    );
-                    let (p_hash, p_root_key, p_aggregate) = cost_return_on_error!(
-                        &mut cost,
-                        primary_merk
-                            .root_hash_key_and_aggregate_data()
-                            .map_err(Error::MerkError)
-                    );
-                    if &p_root_key != primary {
-                        return Err(Error::InvalidInput(
-                            "CountIndexedTree direct insertion: provided \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        // Both roots are Some(_); open and verify they
+                        // match the on-disk state. Mismatch ⇒ the
+                        // element bytes would diverge from on-disk
+                        // state; refuse rather than persist an
+                        // inconsistent root_hash chain.
+                        let child_path_owned = path.derive_owned_with_child(key.to_vec());
+                        let child_path = SubtreePath::from(&child_path_owned);
+                        let primary_merk = cost_return_on_error!(
+                            &mut cost,
+                            self.open_transactional_merk_at_path(
+                                child_path.clone(),
+                                transaction,
+                                Some(batch),
+                                grove_version,
+                            )
+                        );
+                        let (p_hash, p_root_key, p_aggregate) = cost_return_on_error!(
+                            &mut cost,
+                            primary_merk
+                                .root_hash_key_and_aggregate_data()
+                                .map_err(Error::MerkError)
+                        );
+                        if &p_root_key != primary {
+                            return Err(Error::InvalidInput(
+                                "CountIndexedTree direct insertion: provided \
                                  primary_root_key does not match the existing \
                                  primary Merk's root key",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    // The aggregate SHAPE must match the variant too:
-                    // `as_count_u64` reads a count out of both
-                    // `ProvableCount` and `ProvableCountAndProvableSum`, so
-                    // without this a PCIT element could be written over a
-                    // populated PCPSIT primary whose count happens to
-                    // match. The element and the on-disk primary would then
-                    // disagree on arity, and the next `verify_grovedb`
-                    // panics rather than erroring.
-                    if p_aggregate.parent_tree_type() != TreeType::ProvableCountTree {
-                        return Err(Error::InvalidInput(
-                            "CountIndexedTree direct insertion: the existing \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        // The aggregate SHAPE must match the variant too:
+                        // `as_count_u64` reads a count out of both
+                        // `ProvableCount` and `ProvableCountAndProvableSum`, so
+                        // without this a PCIT element could be written over a
+                        // populated PCPSIT primary whose count happens to
+                        // match. The element and the on-disk primary would then
+                        // disagree on arity, and the next `verify_grovedb`
+                        // panics rather than erroring.
+                        if p_aggregate.parent_tree_type() != TreeType::ProvableCountTree {
+                            return Err(Error::InvalidInput(
+                                "CountIndexedTree direct insertion: the existing \
                                  primary Merk is not a provable-count tree; the \
                                  element variant does not match the stored subtree",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    // Fetch the STORED element once — it serves two guards.
-                    //
-                    // In-place conversion guard: a DIFFERENT tree element
-                    // already stored at this key can pass every check above
-                    // (a plain ProvableCountTree's subtree is byte-compatible
-                    // with this variant's primary, so the claimed roots and
-                    // aggregates all match), but the per-axis secondaries
-                    // would start EMPTY over a populated primary — the same
-                    // no-reindex hazard as an axes schema change. Refuse
-                    // the conversion.
-                    //
-                    // Canonical secondary binding (issue #897): the claimed
-                    // secondary_root_key must equal the STORED element's.
-                    // The primary is bound canonically already — its Merk is
-                    // opened with the root key read from the stored parent
-                    // element — but the secondary Merk is opened WITH the
-                    // incoming key, so comparing the returned root key
-                    // against that same input is circular: every row of the
-                    // secondary Merk is a node key in its storage, any of
-                    // them opens successfully as a "root", and the caller
-                    // would commit the hash of an interior/leaf node,
-                    // authenticating a strict subtree of the index as the
-                    // whole index. The stored element's bytes are
-                    // hash-committed by the parent Merk, so its
-                    // secondary_root_key is the canonical authority.
-                    let stored_underlying = cost_return_on_error!(
-                        &mut cost,
-                        Element::get_optional(&subtree_to_insert_into, key, true, grove_version)
-                            .map_err(Error::MerkError)
-                    )
-                    .map(|existing| existing.into_underlying());
-                    match stored_underlying {
-                        Some(Element::ProvableCountIndexedTree(_, stored_secondary, _, _)) => {
-                            if &stored_secondary != secondary {
-                                return Err(Error::InvalidInput(
-                                    "CountIndexedTree direct insertion: provided \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        // Fetch the STORED element once — it serves two guards.
+                        //
+                        // In-place conversion guard: a DIFFERENT tree element
+                        // already stored at this key can pass every check above
+                        // (a plain ProvableCountTree's subtree is byte-compatible
+                        // with this variant's primary, so the claimed roots and
+                        // aggregates all match), but the per-axis secondaries
+                        // would start EMPTY over a populated primary — the same
+                        // no-reindex hazard as an axes schema change. Refuse
+                        // the conversion.
+                        //
+                        // Canonical secondary binding (issue #897): the claimed
+                        // secondary_root_key must equal the STORED element's.
+                        // The primary is bound canonically already — its Merk is
+                        // opened with the root key read from the stored parent
+                        // element — but the secondary Merk is opened WITH the
+                        // incoming key, so comparing the returned root key
+                        // against that same input is circular: every row of the
+                        // secondary Merk is a node key in its storage, any of
+                        // them opens successfully as a "root", and the caller
+                        // would commit the hash of an interior/leaf node,
+                        // authenticating a strict subtree of the index as the
+                        // whole index. The stored element's bytes are
+                        // hash-committed by the parent Merk, so its
+                        // secondary_root_key is the canonical authority.
+                        let stored_underlying = cost_return_on_error!(
+                            &mut cost,
+                            Element::get_optional(subtree_to_insert_into, key, true, grove_version)
+                                .map_err(Error::MerkError)
+                        )
+                        .map(|existing| existing.into_underlying());
+                        match stored_underlying {
+                            Some(Element::ProvableCountIndexedTree(_, stored_secondary, _, _)) => {
+                                if &stored_secondary != secondary {
+                                    return Err(Error::InvalidInput(
+                                        "CountIndexedTree direct insertion: provided \
                                      secondary_root_key does not match the stored element's \
                                      canonical secondary root key",
+                                    ))
+                                    .wrap_with_cost(cost);
+                                }
+                            }
+                            Some(other) if other.is_any_tree() => {
+                                return Err(Error::InvalidInput(
+                                    "CountIndexedTree direct insertion: an existing tree of a \
+                                 different type is stored at this key; converting it in place \
+                                 would leave the secondary index empty over a populated \
+                                 primary (no reindex path)",
+                                ))
+                                .wrap_with_cost(cost);
+                            }
+                            // No stored element / stored non-tree cannot occur:
+                            // `open_transactional_merk_at_path` above only
+                            // succeeds when a stored TREE element exists at this
+                            // key. Fail closed anyway — with no stored
+                            // same-variant element there is no canonical
+                            // authority to bind the claimed secondary to.
+                            _ => {
+                                return Err(Error::InvalidInput(
+                                    "CountIndexedTree direct insertion: a non-empty claim \
+                                 requires an existing stored indexed element to validate \
+                                 the secondary root key against",
                                 ))
                                 .wrap_with_cost(cost);
                             }
                         }
-                        Some(other) if other.is_any_tree() => {
+                        // Also bind the claimed count_value to the primary's
+                        // actual aggregate. Without this, a caller could
+                        // supply correct root keys but a forged count that
+                        // then gets hash-committed and propagated into
+                        // ancestor aggregates.
+                        if p_aggregate.as_count_u64() != *count_value {
                             return Err(Error::InvalidInput(
-                                "CountIndexedTree direct insertion: an existing tree of a \
-                                 different type is stored at this key; converting it in place \
-                                 would leave the secondary index empty over a populated \
-                                 primary (no reindex path)",
-                            ))
-                            .wrap_with_cost(cost);
-                        }
-                        // No stored element / stored non-tree cannot occur:
-                        // `open_transactional_merk_at_path` above only
-                        // succeeds when a stored TREE element exists at this
-                        // key. Fail closed anyway — with no stored
-                        // same-variant element there is no canonical
-                        // authority to bind the claimed secondary to.
-                        _ => {
-                            return Err(Error::InvalidInput(
-                                "CountIndexedTree direct insertion: a non-empty claim \
-                                 requires an existing stored indexed element to validate \
-                                 the secondary root key against",
-                            ))
-                            .wrap_with_cost(cost);
-                        }
-                    }
-                    // Also bind the claimed count_value to the primary's
-                    // actual aggregate. Without this, a caller could
-                    // supply correct root keys but a forged count that
-                    // then gets hash-committed and propagated into
-                    // ancestor aggregates.
-                    if p_aggregate.as_count_u64() != *count_value {
-                        return Err(Error::InvalidInput(
-                            "CountIndexedTree direct insertion: provided \
+                                "CountIndexedTree direct insertion: provided \
                                  count_value does not match the existing \
                                  primary Merk's aggregate count",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    let secondary_merk = cost_return_on_error!(
-                        &mut cost,
-                        self.open_indexed_secondary_at_path(
-                            child_path,
-                            grovedb_element::indexed::IndexAxis::Count,
-                            secondary.clone(),
-                            transaction,
-                            Some(batch),
-                            grove_version,
-                        )
-                    );
-                    let (s_hash, s_root_key, _) = cost_return_on_error!(
-                        &mut cost,
-                        secondary_merk
-                            .root_hash_key_and_aggregate_data()
-                            .map_err(Error::MerkError)
-                    );
-                    if &s_root_key != secondary {
-                        return Err(Error::InvalidInput(
-                            "CountIndexedTree direct insertion: provided \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        let secondary_merk = cost_return_on_error!(
+                            &mut cost,
+                            self.open_indexed_secondary_at_path(
+                                child_path,
+                                grovedb_element::indexed::IndexAxis::Count,
+                                secondary.clone(),
+                                transaction,
+                                Some(batch),
+                                grove_version,
+                            )
+                        );
+                        let (s_hash, s_root_key, _) = cost_return_on_error!(
+                            &mut cost,
+                            secondary_merk
+                                .root_hash_key_and_aggregate_data()
+                                .map_err(Error::MerkError)
+                        );
+                        if &s_root_key != secondary {
+                            return Err(Error::InvalidInput(
+                                "CountIndexedTree direct insertion: provided \
                                  secondary_root_key does not match the existing \
                                  secondary Merk's root key",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    (p_hash, s_hash)
-                };
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        (p_hash, s_hash)
+                    };
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_count_indexed_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         primary_root_hash,
                         secondary_root_hash,
@@ -558,153 +586,151 @@ impl GroveDb {
                 );
             }
             Element::ProvableSumIndexedTree(primary, secondary, sum_value, _) => {
-                let (primary_root_hash, secondary_root_hash) = if primary.is_none()
-                    && secondary.is_none()
-                    && *sum_value == 0
-                {
-                    (NULL_HASH, NULL_HASH)
-                } else {
-                    if primary.is_none() || secondary.is_none() {
-                        return Err(Error::InvalidInput(
-                            "ProvableSumIndexedTree direct insertion: non-empty PSIT must \
+                let (primary_root_hash, secondary_root_hash) =
+                    if primary.is_none() && secondary.is_none() && *sum_value == 0 {
+                        (NULL_HASH, NULL_HASH)
+                    } else {
+                        if primary.is_none() || secondary.is_none() {
+                            return Err(Error::InvalidInput(
+                                "ProvableSumIndexedTree direct insertion: non-empty PSIT must \
                                  have BOTH primary_root_key and secondary_root_key set to \
                                  Some(_); partial state is not permitted",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    let child_path_owned = path.derive_owned_with_child(key.to_vec());
-                    let child_path = SubtreePath::from(&child_path_owned);
-                    let primary_merk = cost_return_on_error!(
-                        &mut cost,
-                        self.open_transactional_merk_at_path(
-                            child_path.clone(),
-                            transaction,
-                            Some(batch),
-                            grove_version,
-                        )
-                    );
-                    let (p_hash, p_root_key, p_aggregate) = cost_return_on_error!(
-                        &mut cost,
-                        primary_merk
-                            .root_hash_key_and_aggregate_data()
-                            .map_err(Error::MerkError)
-                    );
-                    if &p_root_key != primary {
-                        return Err(Error::InvalidInput(
-                            "ProvableSumIndexedTree direct insertion: provided \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        let child_path_owned = path.derive_owned_with_child(key.to_vec());
+                        let child_path = SubtreePath::from(&child_path_owned);
+                        let primary_merk = cost_return_on_error!(
+                            &mut cost,
+                            self.open_transactional_merk_at_path(
+                                child_path.clone(),
+                                transaction,
+                                Some(batch),
+                                grove_version,
+                            )
+                        );
+                        let (p_hash, p_root_key, p_aggregate) = cost_return_on_error!(
+                            &mut cost,
+                            primary_merk
+                                .root_hash_key_and_aggregate_data()
+                                .map_err(Error::MerkError)
+                        );
+                        if &p_root_key != primary {
+                            return Err(Error::InvalidInput(
+                                "ProvableSumIndexedTree direct insertion: provided \
                                  primary_root_key does not match the existing primary Merk's \
                                  root key",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    // Bind the claimed sum_value to the primary's actual
-                    // aggregate so a forged sum cannot be hash-committed
-                    // and propagated into ancestor aggregates.
-                    if p_aggregate.parent_tree_type() != TreeType::ProvableSumTree {
-                        return Err(Error::InvalidInput(
-                            "ProvableSumIndexedTree direct insertion: the \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        // Bind the claimed sum_value to the primary's actual
+                        // aggregate so a forged sum cannot be hash-committed
+                        // and propagated into ancestor aggregates.
+                        if p_aggregate.parent_tree_type() != TreeType::ProvableSumTree {
+                            return Err(Error::InvalidInput(
+                                "ProvableSumIndexedTree direct insertion: the \
                                  existing primary Merk is not a provable-sum \
                                  tree; the element variant does not match the \
                                  stored subtree",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    // Fetch the STORED element once — it serves two guards.
-                    //
-                    // In-place conversion guard: a DIFFERENT tree element
-                    // already stored at this key can pass every check above
-                    // (a plain ProvableSumTree's subtree is byte-compatible
-                    // with this variant's primary, so the claimed roots and
-                    // aggregates all match), but the per-axis secondaries
-                    // would start EMPTY over a populated primary — the same
-                    // no-reindex hazard as an axes schema change. Refuse
-                    // the conversion.
-                    //
-                    // Canonical secondary binding (issue #897): the claimed
-                    // secondary_root_key must equal the STORED element's —
-                    // the secondary Merk below is opened WITH the incoming
-                    // key, so comparing the returned root key against that
-                    // same input is circular (any row node key of the
-                    // secondary opens successfully as a "root"). See the
-                    // PCIT arm above for the full rationale.
-                    let stored_underlying = cost_return_on_error!(
-                        &mut cost,
-                        Element::get_optional(&subtree_to_insert_into, key, true, grove_version)
-                            .map_err(Error::MerkError)
-                    )
-                    .map(|existing| existing.into_underlying());
-                    match stored_underlying {
-                        Some(Element::ProvableSumIndexedTree(_, stored_secondary, _, _)) => {
-                            if &stored_secondary != secondary {
-                                return Err(Error::InvalidInput(
-                                    "ProvableSumIndexedTree direct insertion: provided \
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        // Fetch the STORED element once — it serves two guards.
+                        //
+                        // In-place conversion guard: a DIFFERENT tree element
+                        // already stored at this key can pass every check above
+                        // (a plain ProvableSumTree's subtree is byte-compatible
+                        // with this variant's primary, so the claimed roots and
+                        // aggregates all match), but the per-axis secondaries
+                        // would start EMPTY over a populated primary — the same
+                        // no-reindex hazard as an axes schema change. Refuse
+                        // the conversion.
+                        //
+                        // Canonical secondary binding (issue #897): the claimed
+                        // secondary_root_key must equal the STORED element's —
+                        // the secondary Merk below is opened WITH the incoming
+                        // key, so comparing the returned root key against that
+                        // same input is circular (any row node key of the
+                        // secondary opens successfully as a "root"). See the
+                        // PCIT arm above for the full rationale.
+                        let stored_underlying = cost_return_on_error!(
+                            &mut cost,
+                            Element::get_optional(subtree_to_insert_into, key, true, grove_version)
+                                .map_err(Error::MerkError)
+                        )
+                        .map(|existing| existing.into_underlying());
+                        match stored_underlying {
+                            Some(Element::ProvableSumIndexedTree(_, stored_secondary, _, _)) => {
+                                if &stored_secondary != secondary {
+                                    return Err(Error::InvalidInput(
+                                        "ProvableSumIndexedTree direct insertion: provided \
                                      secondary_root_key does not match the stored element's \
                                      canonical secondary root key",
-                                ))
-                                .wrap_with_cost(cost);
+                                    ))
+                                    .wrap_with_cost(cost);
+                                }
                             }
-                        }
-                        Some(other) if other.is_any_tree() => {
-                            return Err(Error::InvalidInput(
+                            Some(other) if other.is_any_tree() => {
+                                return Err(Error::InvalidInput(
                                 "ProvableSumIndexedTree direct insertion: an existing tree of a \
                                  different type is stored at this key; converting it in place \
                                  would leave the secondary index empty over a populated \
                                  primary (no reindex path)",
                             ))
                             .wrap_with_cost(cost);
-                        }
-                        // No stored element / stored non-tree cannot occur:
-                        // `open_transactional_merk_at_path` above only
-                        // succeeds when a stored TREE element exists at this
-                        // key. Fail closed anyway.
-                        _ => {
-                            return Err(Error::InvalidInput(
-                                "ProvableSumIndexedTree direct insertion: a non-empty claim \
+                            }
+                            // No stored element / stored non-tree cannot occur:
+                            // `open_transactional_merk_at_path` above only
+                            // succeeds when a stored TREE element exists at this
+                            // key. Fail closed anyway.
+                            _ => {
+                                return Err(Error::InvalidInput(
+                                    "ProvableSumIndexedTree direct insertion: a non-empty claim \
                                  requires an existing stored indexed element to validate the \
                                  secondary root key against",
+                                ))
+                                .wrap_with_cost(cost);
+                            }
+                        }
+                        if p_aggregate.as_sum_i64() != *sum_value {
+                            return Err(Error::InvalidInput(
+                                "ProvableSumIndexedTree direct insertion: provided \
+                                 sum_value does not match the existing primary Merk's \
+                                 aggregate sum",
                             ))
                             .wrap_with_cost(cost);
                         }
-                    }
-                    if p_aggregate.as_sum_i64() != *sum_value {
-                        return Err(Error::InvalidInput(
-                            "ProvableSumIndexedTree direct insertion: provided \
-                                 sum_value does not match the existing primary Merk's \
-                                 aggregate sum",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    let secondary_merk = cost_return_on_error!(
-                        &mut cost,
-                        self.open_indexed_secondary_at_path(
-                            child_path,
-                            grovedb_element::indexed::IndexAxis::Sum,
-                            secondary.clone(),
-                            transaction,
-                            Some(batch),
-                            grove_version,
-                        )
-                    );
-                    let (s_hash, s_root_key, _) = cost_return_on_error!(
-                        &mut cost,
-                        secondary_merk
-                            .root_hash_key_and_aggregate_data()
-                            .map_err(Error::MerkError)
-                    );
-                    if &s_root_key != secondary {
-                        return Err(Error::InvalidInput(
-                            "ProvableSumIndexedTree direct insertion: provided \
+                        let secondary_merk = cost_return_on_error!(
+                            &mut cost,
+                            self.open_indexed_secondary_at_path(
+                                child_path,
+                                grovedb_element::indexed::IndexAxis::Sum,
+                                secondary.clone(),
+                                transaction,
+                                Some(batch),
+                                grove_version,
+                            )
+                        );
+                        let (s_hash, s_root_key, _) = cost_return_on_error!(
+                            &mut cost,
+                            secondary_merk
+                                .root_hash_key_and_aggregate_data()
+                                .map_err(Error::MerkError)
+                        );
+                        if &s_root_key != secondary {
+                            return Err(Error::InvalidInput(
+                                "ProvableSumIndexedTree direct insertion: provided \
                                  secondary_root_key does not match the existing secondary \
                                  Merk's root key",
-                        ))
-                        .wrap_with_cost(cost);
-                    }
-                    (p_hash, s_hash)
-                };
+                            ))
+                            .wrap_with_cost(cost);
+                        }
+                        (p_hash, s_hash)
+                    };
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_count_indexed_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         primary_root_hash,
                         secondary_root_hash,
@@ -831,7 +857,7 @@ impl GroveDb {
                     // PCIT arm above for the full rationale.
                     let stored_underlying = cost_return_on_error!(
                         &mut cost,
-                        Element::get_optional(&subtree_to_insert_into, key, true, grove_version)
+                        Element::get_optional(subtree_to_insert_into, key, true, grove_version)
                             .map_err(Error::MerkError)
                     )
                     .map(|existing| existing.into_underlying());
@@ -938,7 +964,7 @@ impl GroveDb {
                 cost_return_on_error_into!(
                     &mut cost,
                     element.insert_count_indexed_subtree(
-                        &mut subtree_to_insert_into,
+                        subtree_to_insert_into,
                         key,
                         primary_root_hash,
                         second_hash,
@@ -959,6 +985,6 @@ impl GroveDb {
             }
         }
 
-        Ok(subtree_to_insert_into).wrap_with_cost(cost)
+        Ok(()).wrap_with_cost(cost)
     }
 }
