+++ grovedb/src/operations/insert/insert_on_transaction/v1.rs
@@ -1,34 +1,14 @@
-//! `insert_on_transaction` — **v1** (`GROVE_V4`+).
+//! Automatic backward-reference maintenance on GROVE_V4.
 //!
-//! A behaviour-preserving router. A call that neither inserts a
-//! [`Element::BidirectionalReference`] nor sets
-//! [`InsertOptions::backward_references_policy`] runs the exact shipped
-//! flow ([`super::v0::insert_on_transaction_body`]) — identical root hashes
-//! and costs. The rest run through the `MerkCache`-based flow below, which
-//! keeps every touched subtree open in one cache so backward-reference
-//! bookkeeping, chain propagation, and ordinary parent propagation all see
-//! each other's uncommitted writes (see `adr/bidirectional_references.md`
-//! and `adr/merk_cache.md`).
-//!
-//! ## Element support under the backward-references flow
-//!
-//! The `MerkCache` flow supports items, references (all three variants),
-//! and empty plain-Merk trees. Specialized tree types (commitment / MMR /
-//! bulk-append / dense / private document store / indexed trees) and the
-//! aggregation wrappers are rejected when `backward_references_policy`
-//! is set — their child-hash conventions live in
-//! `add_element_on_transaction` and have no backward-references semantics
-//! yet. Insert them without the flag (they cannot be targeted by
-//! bidirectional references anyway).
+//! Prepare the mutated node through the old-value observer and retain its
+//! traversal in the same Merk used by the write. Reference maintenance and
+//! parent propagation share that cache until the atomic batch is complete.
+
+use grovedb_merk::element::tree_type::ElementTreeTypeExtensions;
+use std::collections::HashMap;
 
 use grovedb_costs::{cost_return_on_error, cost_return_on_error_no_add, CostResult, CostsExt};
-use grovedb_merk::{
-    element::{
-        costs::ElementCostExtensions, get::ElementFetchFromStorageExtensions,
-        insert::ElementInsertToStorageExtensions,
-    },
-    tree::NULL_HASH,
-};
+use grovedb_merk::element::{costs::ElementCostExtensions, insert::Delta};
 use grovedb_path::SubtreePath;
 use grovedb_storage::StorageBatch;
 use grovedb_version::version::GroveVersion;
@@ -36,11 +16,10 @@ use grovedb_version::version::GroveVersion;
 use super::super::InsertOptions;
 use crate::{
     bidirectional_references::{
-        basic_sectioned_removal, process_bidirectional_reference_insertion,
-        process_update_element_with_backward_references,
+        basic_sectioned_removal, check_carried_referrers_fit,
+        process_bidirectional_reference_insertion, process_update_element_with_backward_references,
     },
     merk_cache::MerkCache,
-    reference_path::follow_reference,
     Element, Error, GroveDb, Transaction,
 };
 
@@ -48,33 +27,23 @@ pub(super) fn insert_on_transaction<'db, 'b, B: AsRef<[u8]>>(
     db: &GroveDb,
     path: SubtreePath<'b, B>,
     key: &[u8],
-    element: Element,
+    mut element: Element,
     options: InsertOptions,
     transaction: &'db Transaction,
     batch: &StorageBatch,
     grove_version: &GroveVersion,
 ) -> CostResult<(), Error> {
-    // Backward-references elements are never wrapped: the wrappers'
-    // constructors refuse them and deserialization rejects the shape, so a
-    // hand-built `NonCounted(family)` must not reach either route below
-    // (the plain route would store bytes no reader accepts).
     if element.is_wrapped() && element.underlying().supports_backward_references() {
         return Err(Error::InvalidInput(
-            "backward-references elements (BidirectionalReference, \
-             ItemWithBackwardsReferences, SumItemWithBackwardsReferences, \
-             ItemWithSumItemWithBackwardsReferences) cannot be wrapped in NonCounted / \
-             NotSummed / NotCountedOrSummed",
+            "backward-references elements cannot be wrapped",
         ))
         .wrap_with_cost(Default::default());
     }
-
-    // A bidirectional reference must always register itself in its target's
-    // meta storage, flag or no flag; everything else opts into the
-    // backward-references flow via the flag.
-    if matches!(element, Element::BidirectionalReference(..))
-        || options.backward_references_policy.maintains()
+    // A new bidirectional reference must register its target even under Skip.
+    if !options.backward_references_policy.maintains()
+        && !matches!(element, Element::BidirectionalReference(..))
     {
-        insert_with_backward_references(
+        return super::v0::insert_on_transaction_body(
             db,
             path,
             key,
@@ -83,263 +52,186 @@ pub(super) fn insert_on_transaction<'db, 'b, B: AsRef<[u8]>>(
             transaction,
             batch,
             grove_version,
-        )
-    } else {
-        super::v0::insert_on_transaction_body(
-            db,
-            path,
-            key,
-            element,
-            options,
-            transaction,
-            batch,
-            grove_version,
-        )
+        );
     }
-}
-
-/// The `MerkCache`-based insert flow with backward-references bookkeeping.
-fn insert_with_backward_references<'db, 'b, B: AsRef<[u8]>>(
-    db: &GroveDb,
-    path: SubtreePath<'b, B>,
-    key: &[u8],
-    element: Element,
-    options: InsertOptions,
-    transaction: &'db Transaction,
-    batch: &StorageBatch,
-    grove_version: &GroveVersion,
-) -> CostResult<(), Error> {
+    if grove_version
+        .grovedb_versions
+        .operations
+        .insert
+        .add_element_on_transaction
+        != 2
+    {
+        return super::v0::insert_on_transaction_body(
+            db,
+            path,
+            key,
+            element,
+            options,
+            transaction,
+            batch,
+            grove_version,
+        );
+    }
     let mut cost = Default::default();
-
-    let cache = MerkCache::new(db, transaction, grove_version);
-
-    let mut subtree_to_insert_into =
-        cost_return_on_error!(&mut cost, cache.get_merk(path.derive_owned()));
-
-    if options.checks_for_override() {
-        let maybe_element_bytes = cost_return_on_error!(
-            &mut cost,
-            subtree_to_insert_into.for_merk(|m| m
-                .get(
+    let mut merk = cost_return_on_error!(
+        &mut cost,
+        db.open_transactional_merk_at_path(path.clone(), transaction, Some(batch), grove_version)
+    );
+    let previous = cost_return_on_error!(
+        &mut cost,
+        (|| {
+            let mut cost = Default::default();
+            cost_return_on_error_no_add!(
+                cost,
+                crate::operations::indexed_tree::reject_generic_write_into_indexed_primary(
+                    merk.tree_type,
+                    "insert",
+                )
+            );
+            let mut previous = None;
+            cost_return_on_error!(
+                &mut cost,
+                merk.observe_old_value(
                     key,
-                    true,
                     Some(&Element::value_defined_cost_for_serialized_value),
+                    &mut |_, bytes| {
+                        previous = Some(Element::deserialize(bytes, grove_version));
+                    },
                     grove_version,
                 )
-                .map_err(|e| Error::CorruptedData(e.to_string())))
+                .map_err(Error::MerkError)
+            );
+            previous
+                .transpose()
+                .map_err(Error::from)
+                .wrap_with_cost(cost)
+        })()
+    );
+    if previous.is_some() && options.validate_insertion_does_not_override {
+        return Err(Error::OverrideNotAllowed(
+            "insertion not allowed to override",
+        ))
+        .wrap_with_cost(cost);
+    }
+    if previous.as_ref().is_some_and(Element::is_any_tree)
+        && options.validate_insertion_does_not_override_tree
+    {
+        return Err(Error::OverrideNotAllowed(
+            "insertion not allowed to override tree",
+        ))
+        .wrap_with_cost(cost);
+    }
+    if previous.as_ref().is_some_and(|old| {
+        old.is_any_tree()
+            && !old.uses_non_merk_data_storage()
+            && old
+                .root_key_and_tree_type()
+                .is_some_and(|(root, _)| root.is_some())
+    }) && !cost_return_on_error!(
+        &mut cost,
+        db.backward_reference_participants(
+            &path.derive_owned_with_child(key).to_vec(),
+            transaction,
+            grove_version,
+        )
+    )
+    .is_empty()
+    {
+        return Err(Error::NotSupported(
+            "delete a subtree containing backward-reference participants before replacing it"
+                .to_owned(),
+        ))
+        .wrap_with_cost(cost);
+    }
+    let needs_maintenance = element.supports_backward_references()
+        || previous
+            .as_ref()
+            .is_some_and(Element::supports_backward_references);
+    if !needs_maintenance {
+        cost_return_on_error!(
+            &mut cost,
+            db.add_element_to_cached_merk_v2(
+                &mut merk,
+                path.clone(),
+                key,
+                element,
+                options,
+                transaction,
+                batch,
+                grove_version,
+            )
         );
-        if let Some(element_bytes) = maybe_element_bytes {
-            if options.validate_insertion_does_not_override {
-                return Err(Error::OverrideNotAllowed(
-                    "insertion not allowed to override",
-                ))
-                .wrap_with_cost(cost);
-            }
-            if options.validate_insertion_does_not_override_tree {
-                let element = cost_return_on_error_no_add!(
-                    cost,
-                    Element::deserialize(element_bytes.as_slice(), grove_version).map_err(|_| {
-                        Error::CorruptedData(String::from("unable to deserialize element"))
-                    })
-                );
-                if element.is_any_tree() {
-                    return Err(Error::OverrideNotAllowed(
-                        "insertion not allowed to override tree",
-                    ))
-                    .wrap_with_cost(cost);
-                }
-            }
-        }
+        let mut merks = HashMap::new();
+        merks.insert(path.clone(), merk);
+        cost_return_on_error!(
+            &mut cost,
+            db.propagate_changes_with_transaction(merks, path, transaction, batch, grove_version,)
+        );
+        return Ok(()).wrap_with_cost(cost);
     }
+    let cache = MerkCache::with_prepared_merk(
+        db,
+        transaction,
+        grove_version,
+        path.derive_owned(),
+        merk,
+        batch,
+    );
+    let mut handle = cost_return_on_error!(&mut cost, cache.get_merk(path.derive_owned()));
 
-    match element {
-        Element::BidirectionalReference(reference, flags) => {
+    if let Element::BidirectionalReference(reference, flags) = element {
+        cost_return_on_error!(
+            &mut cost,
+            process_bidirectional_reference_insertion(
+                &cache,
+                path,
+                key,
+                reference,
+                flags,
+                Some(options),
+            )
+        );
+    } else {
+        if let Some(refs) = element.backward_references_mut() {
+            *refs = previous
+                .as_ref()
+                .and_then(Element::backward_references)
+                .map(<[_]>::to_vec)
+                .unwrap_or_default();
+            cost_return_on_error_no_add!(cost, check_carried_referrers_fit(&element));
+        }
+        let unchanged_family =
+            element.supports_backward_references() && previous.as_ref() == Some(&element);
+        if !unchanged_family {
             cost_return_on_error!(
                 &mut cost,
-                process_bidirectional_reference_insertion(
-                    &cache,
-                    path,
+                handle.for_merk(|merk| db.add_element_to_cached_merk_v2(
+                    merk,
+                    path.clone(),
                     key,
-                    reference,
-                    flags,
-                    Some(options)
-                )
-            );
-        }
-        Element::Reference(ref reference_path, ..)
-        | Element::ReferenceWithSumItem(ref reference_path, ..) => {
-            let resolved_reference = cost_return_on_error!(
-                &mut cost,
-                follow_reference(&cache, path.derive_owned(), key, reference_path.clone())
-            );
-            let referenced_item: Element = resolved_reference.target_element;
-
-            if referenced_item.is_any_tree() {
-                return Err(Error::NotSupported(
-                    "References cannot point to subtrees".to_owned(),
+                    element.clone(),
+                    options,
+                    transaction,
+                    batch,
+                    grove_version,
                 ))
-                .wrap_with_cost(cost);
-            }
-
-            let delta = cost_return_on_error!(
-                &mut cost,
-                subtree_to_insert_into.for_merk(|m| {
-                    element
-                        .insert_reference_if_changed_value(
-                            m,
-                            key,
-                            resolved_reference.target_node_value_hash,
-                            Some(options.as_merk_options()),
-                            grove_version,
-                        )
-                        .map_err(Error::MerkError)
-                })
-            );
-
-            cost_return_on_error!(
-                &mut cost,
-                process_update_element_with_backward_references(
-                    &cache,
-                    subtree_to_insert_into.clone(),
-                    path.derive_owned(),
-                    key,
-                    delta,
-                    &mut basic_sectioned_removal()
-                )
-            );
-        }
-        Element::Tree(ref root_key, _)
-        | Element::SumTree(ref root_key, ..)
-        | Element::BigSumTree(ref root_key, ..)
-        | Element::CountTree(ref root_key, ..)
-        | Element::CountSumTree(ref root_key, ..)
-        | Element::ProvableCountTree(ref root_key, ..)
-        | Element::ProvableCountSumTree(ref root_key, ..)
-        | Element::ProvableSumTree(ref root_key, ..)
-        | Element::ProvableCountProvableSumTree(ref root_key, ..) => {
-            if root_key.is_some() {
-                return Err(Error::InvalidCodeExecution(
-                    "a tree should be empty at the moment of insertion when not using batches",
-                ))
-                .wrap_with_cost(cost);
-            }
-            let delta = cost_return_on_error!(
-                &mut cost,
-                subtree_to_insert_into.for_merk(|m| {
-                    element
-                        .insert_subtree_if_changed(
-                            m,
-                            key,
-                            NULL_HASH,
-                            Some(options.as_merk_options()),
-                            grove_version,
-                        )
-                        .map_err(Error::MerkError)
-                })
-            );
-
-            cost_return_on_error!(
-                &mut cost,
-                process_update_element_with_backward_references(
-                    &cache,
-                    subtree_to_insert_into.clone(),
-                    path.derive_owned(),
-                    key,
-                    delta,
-                    &mut basic_sectioned_removal()
-                )
-            );
-        }
-        Element::CommitmentTree(..)
-        | Element::MmrTree(..)
-        | Element::BulkAppendTree(..)
-        | Element::DenseAppendOnlyFixedSizeTree(..)
-        | Element::PrivateDocumentStore(..)
-        | Element::ProvableSumIndexedTree(..)
-        | Element::ProvableCountIndexedTree(..)
-        | Element::ProvableCountProvableSumIndexedTree(..)
-        | Element::NonCounted(..)
-        | Element::NotSummed(..)
-        | Element::NotCountedOrSummed(..) => {
-            // These carry child-hash conventions or wrapper semantics that
-            // the MerkCache flow does not model; and none of them can be
-            // targeted by bidirectional references. Insert them without the
-            // flag.
-            return Err(Error::NotSupported(
-                "this element type cannot be inserted with backward_references_policy set; \
-                 insert it without the flag"
-                    .to_owned(),
-            ))
-            .wrap_with_cost(cost);
-        }
-        Element::Item(..)
-        | Element::SumItem(..)
-        | Element::ItemWithSumItem(..)
-        | Element::ItemWithBackwardsReferences(..)
-        | Element::SumItemWithBackwardsReferences(..)
-        | Element::ItemWithSumItemWithBackwardsReferences(..) => {
-            // A backward-references element's referrer list is bookkeeping
-            // this flow maintains: carry the stored list over onto the new
-            // element so an update never silently drops registrations (and
-            // so the changed/unchanged comparison reflects the LOGICAL
-            // value, the referrer lists being equal on both sides).
-            let mut element = element;
-            if element.supports_backward_references() {
-                let previous = cost_return_on_error!(
-                    &mut cost,
-                    subtree_to_insert_into.for_merk(|m| {
-                        Element::get_optional(m, key, true, grove_version).map_err(Error::MerkError)
-                    })
-                );
-                if let Some(refs) = element.backward_references_mut() {
-                    // The stored list is authoritative; whatever the caller
-                    // supplied is not theirs to claim — forged entries would
-                    // later let cascades and propagations follow arbitrary
-                    // inverted paths.
-                    *refs = previous
-                        .as_ref()
-                        .and_then(|p| p.backward_references())
-                        .map(|p| p.to_vec())
-                        .unwrap_or_default();
-                }
-                // The carried-over referrers must fit the capacity the new
-                // element declares (checked before the write so the refusal
-                // is the family rule, not a serialization failure).
-                cost_return_on_error_no_add!(
-                    cost,
-                    crate::bidirectional_references::check_carried_referrers_fit(&element)
-                );
-            }
-            let delta = cost_return_on_error!(
-                &mut cost,
-                subtree_to_insert_into.for_merk(|m| {
-                    element
-                        .insert_if_changed_value(
-                            m,
-                            key,
-                            Some(options.as_merk_options()),
-                            grove_version,
-                        )
-                        .map_err(Error::MerkError)
-                })
             );
             cost_return_on_error!(
                 &mut cost,
                 process_update_element_with_backward_references(
                     &cache,
-                    subtree_to_insert_into.clone(),
+                    handle,
                     path.derive_owned(),
                     key,
-                    delta,
-                    &mut basic_sectioned_removal()
+                    Delta {
+                        new: Some(&element),
+                        old: previous
+                    },
+                    &mut basic_sectioned_removal(),
                 )
             );
         }
     }
-
-    let result_batch = cost_return_on_error!(&mut cost, cache.into_batch());
-
-    batch.merge(*result_batch);
-
+    cost_return_on_error!(&mut cost, cache.finish_prepared());
     Ok(()).wrap_with_cost(cost)
 }
